requirements.txt
fastapi
uvicorn
python-dotenv
jinja2
websockets
deepgram-sdk>=3.0.0
